Add new log_manager role which can update the S3 bucket policy
Users assuming this role can edit the S3 bucket policy which lists the accounts
able to ship CloudTrail logs into the bucket.

Introduces a new "log_managers" variable which can be used to list ARNs of the
IAM users which can assume this role.

Add ListLogGroupTags access is in a separate statement as the ReadOnly policy
from AWS doesn't seem to include this. We should raise a support ticket for them
to update it.

With this policy Terraform is able to update the CloudTrails S3 policy document.